Safari-609.1.11.1/LayoutTests/fast/dom/StyleSheet/gc-inline-style-cssvalues.html

<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body style="color:black">
<script>
    description("Test that custom properties on CSSValues of inline style declarations survive GC");

    var expr = 'document.body.style.getPropertyCSSValue("color")';
    shouldBe(expr + ".cssText", '"black"');
    eval(expr).foo = "bar"
    gc();
    shouldBe(expr + ".foo", "'bar'");

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
