safari-534.53-branch/LayoutTests/fast/dom/StyleSheet/gc-inline-style-cssvalues-expected.txt

Test that custom properties on CSSValues of inline style declarations survive GC

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.body.style.getPropertyCSSValue("color").cssText is "black"
PASS document.body.style.getPropertyCSSValue("color").foo is 'bar'
PASS successfullyParsed is true

TEST COMPLETE
